Introduce new action intent ALARM_DISMISS (not used yet though).
Moved handling of shared state (i.e. setting an alarm and handling the currently set alarm) to the RetroTimer class, and made it an Application.

Check on boot and on app start-up if any alarms are pending (or have passed), and handle it accordingly (i.e. set the alarm or clean up after it, respectively).

Added some more permissions from the AlarmClock app.