Fixed bug of cancelled alarm after locking screen with app open
If you set an alarm and locked the screen (or let it time out), then when the alarm triggers and starts sounding, the TimerSet activity will be brought to foreground for a split second before the TimerAlert activity is launched.

Unfortunately, the TimerSet activity used to dismiss any currently sounding alarms from its onResume(), causing the alarm sound to be dismissed almost immediately after starting.

Now, the alarm is not dismissed directly when the TimerSet activity is resumed, but only if the user interacts with the activity (e.g. tries to set a new alarm).