Draw numbers on dial even with HW acceleration turned on
The drawTextOnPath() method is not supported with HW acceleration turned on (see the [Android API Guides][1]), so if you are using a recent Android version and have turned on "Force GPU rendering" (under Settings, Developer options), the numbers will not be visible.

This problem is fixed by disabling HW acceleration for the TimerView.

This commit fixes GitHub issue #8.

[1] http://developer.android.com/guide/topics/graphics/hardware-accel.html