Add code to detect missed timeouts
Check, when we can, if the time left to alert has gone negative. This should never happen, but seems to happen to some people (due to running task killers, perhaps?). I have not been able to reproduce it.

If it happens, we try to return the app to it's normal state, and then throw a RuntimeException to let the user send a crash report.